Rename managebuffer view to cardlist view
